Test database setup

+++ code/server/server_test/test_database_setup.py
@@ -6,6 +6,11 @@ from trends.db.admin import article_admin
 from trends.db.article import Article
 
 
+confirmation = input("This will delete all database entries and reset it to the default test database. Are you sure? "
+                     "(Y/N)")
+if confirmation.lower() != "y":
+    exit(0)
+
 Article.objects.all().delete()
 
 with open('/article_html/1.html', 'w') as f:
